Remove redundant extern keywords (#1050)
It is written in the "6.2.2 Linkages of identifiers" chapter of C99 standard
(https://www.open-std.org/jtc1/sc22/wg14/www/docs/n1256.pdf):
"If the declaration of an identifier for a function has no storage-class
specifier, its linkage is determined exactly as if it were declared with
the storage-class specifier extern". The "6.9 External definitions" chapter
contains the following: "An external definition is an external declaration that
is also a definition of a function (other than an inline definition) or
an object". So the extern keyword is redundant in function definitions.

Greenplum specific code is changed only.