Treat S3 errors as query errors (#1097)
It is possible for reader_cleanup and writer_cleanup to produce errors (via
S3Exception), resulting in incorrect query completion. Errors in these
functions should be treated as ERRORs, not WARNINGs.

However, we cannot throw ERRORs inside resource owner callbacks, so during an
actual abort we still use WARNINGs. This shouldn't be an issue because such an
abort should only occur in erroneous conditions anyway.

Tests are not presented, as there is no existing test infrastructure to make an
auto test with S3 storage.

Ticket: ADBDEV-6588