Freeze LLVM to 17 version in Dockerfile (#1113)
In our rockylinux8 docker container Clang is installed like dnf install clang
which doesn’t force any version and uses that version which is considered
current in the repo. Until now, the default LLVM version was 17, but it has
been recently promoted to 18 which caused our JIT tests to fail.

The reason of this problem is that postgres isn’t ready for LLVM 18, that is,
it uses an old API which isn’t compatible with new LLVM 18. Currently, there's
no patch even in the postgres upstream which addresses this problem.

As a temporary solution, this patch freezes Clang and LLVM packages to the 17
version

Ticket: ADBDEV-6677

Co-authored-by: Georgy Shelkovy <g.shelkovy@arenadata.io>