Don't overwrite gpexpand template on the same host (#1100)
In RedOS, when both the master and a segment are on the same host, running
gpexpand correctly was impossible.

Typically, the master is excluded from the gpexpand host array. However, if both
the master and a segment are on the same host, the master's host gets included
in the array. The issue arises due to the scp command corrupting and truncating
the gpexpand_schema.tar file when the host, source and destination paths are the
same.

Don't issue a scp command if we are dealing with the local host with and paths
are the same. Note that gpexpand always uses relative filepaths for scp's
destination, and that they are relative to /home/gpadmin/.

This problem does not occur in GPDB 7, as it uses rsync instead of scp, which
does not have said issue.

Ticket: ADBDEV-6472