Ensure proper order of statements in fts_segment_reset test (#1099)
Ensure proper order of statements in the test fts_segment_reset.

The aforementioned isolation test contains statements that are started in
parallel at the same time. Under some circumstances, e.g., when using a Docker
container based on Ubuntu 22.04, these statements are executed in an unexpected
order, causing the test to fail.

This commit adds a fault injector to enforce the order of statements by
introducing a dependency between getting segment master into reset and
the start of next command.