Fix initplan functions returning NULL (#1141)
Previously in some rare circumstances, specifically if an initplan function
was contained inside a non-trivial projection node during planning, and in a
query that can't be optimized further like `INSERT INTO table SELECT ...`,
then `FunctionScan` node was copied into the initplan before its `targetList`
was initialized. This caused `FunctionScan` inside initplans to have empty
`targetList`, which resulted in that function always producing NULLs. For
example, the following query would produce this issue: `INSERT INTO table
SELECT a + 1 AS a, b FROM initplan_func()`, assuming `initplan_func()` is
marked as ON INITPLAN and returns a table with `a` and `b` attributes.

This was caused by an optimization during planning when projection nodes would
skip building the target list of some subplans via `CP_IGNORE_TLIST` flag,
because they are later overwritten by the projection. This patch adds a check
that disables `CP_IGNORE_TLIST` specifically for FunctionScans executing on
initplan. This way the original `targetList` of FunctionScan persists in the
initplan's copy, so the resulting tuples are stored into the tuplestore in
their native projection. This is acceptable because the main FunctionScan then
transforms it into the required projection when it reads the tuplestore.