Check collector has started in pg_wait_sampling_reset_profile.
The worker might have not started yet or it may never start,
because its registration was cancelled due to worker limit.

This commit adds a check for NULL value of pgws_collector_hdr->latch.
The previous usage in pg_wait_sampling.c has such a check, we
should do the same here.