Fix incorrect behavior of gp_percentile_* functions (#1846)
Commit b90f952 added five GPDB-specific functions
gp_percentile_cont_float8_transition, gp_percentile_cont_interval_transition,
gp_percentile_cont_timestamp_transition,
gp_percentile_cont_timestamptz_transition and gp_percentile_disc_transition.

1) All these five functions use five arguments internally in the C
implementation, but they were added to the catalog in the file
src/include/catalog/pg_proc.dat with four arguments, which could lead to
incorrect results, assertions when freeing a null pointer and even segfaults
when dereferencing a null pointer. This patch adds the missing arguments.

2) All these five functions under certain conditions can return their first
argument, but if the first argument was NULL (and the NULL argument is passed
with the additional isnull flag), then when it was returned, all these five
functions lost the additional isnull flag, which could also lead to incorrect
results (for the non-varlena types of the first two arguments), and to
assertions when freeing a null pointer and even segfaults when dereferencing a
null pointer (for the varlena types). This patch adds saving of the additional
isnull flag, and now under the same conditions, all these five functions return
NULL.

The patch adds three tests for each of the five functions.

Ticket: ADBDEV-7770