enable isolation2 test: parallel_retrieve_cursor/status_check
It was disabled due to unstable output. Sometimes test was failed, because error
after creating parallel retrieve cursor was printed immediately, but should be
printed, when "rollback" is executed.

Flakiness cased, because session was not disconnected at the previous test. But
the previous test generates "query finish request to process <PID>", which is
also sent to session of the next test. It lead to immediately processing all
errors, which are generated be cursors. But at the normal case errors, should be
processing at the end of transaction.

Patch adds quitting for all connections to the previous test. After that, the
previous test does not affect the next test.