fix ADBDEV-8207: Flaky fts_segment_reset test
In some cases processing of injected panic fault can take a long time.
So queries, sent after sending fault, can either complete
successfully or end with an error.
To prevent this uncertainty the waiting mechanism was introduced in
test - before sending next commands after injecting panic fault, we wait
for it to be fully processed that is to trigger SIGCHLD.