Implement skeleton for basic shrink scenario
This it the rebalance feature migration commit.
Starting from this patch we exclude the early MVP code of gprebalance.
Thus, files gprebalance, rebalance_validator.py, rebalance.py are deleted.
Original commit message:

This patch implements a state machine skeleton for a basic shrink scenario based
on 'transitions' library. It consists of a new 'ggrebalance' tool, which will be
a single entry point for shrink, expand, and cluster rebalance functionality,
and 'shrink.py', which contains the state machine itself with the shrink logic.

The main purpose of this half-MVP is to evaluate the state machine pattern
suitability. Therefore it implements only a limited set of requirements for the
shrink, which allows you to support basic shrink workflow.

(cherry picked from commit arenadata/gpdb@69fb3e9)