[ORCA] Fix infinite recursion during statistic derivation
During the exploration phase, new groups of equivalent expressions are created.
In this process, some groups are marked as duplicates. After exploration,
expressions from duplicate groups are moved into the group they duplicate.

In cases where a duplicate group contains an expression that references the
duplicated group, merging them results in a situation where a group contains an
expression that references the very group it belongs to. This leads to infinite
recursion during statistics derivation.

The fix is to improve the cycle-detection logic so that it can recognize when an
expression references the group it resides in.

(cherry picked from commit b5bdddabcd5e517948bbf23e9e0f039d62df05ba)