Update limit path creation
The limit path had unnecessary checks for volatility, so in cases of
creating limit over general or segmentGeneral locuses (replicated
tables or generate_series() for example) limits with just constants
wouldn't pass this checks and generate right path.
There's no need for limit to have volatile function inside of it
to force execution on one segment because results of such limit path
    ->  Limit
        ->  Seq Scan on t_repl
is also undetermined in case of execution on each segment (as there is
no guarantees on how seqscan will return it's rows).