Fix the code for compatibility with PyGreSQL 5.2.5
The following points have been fixed:
1. PyGreSQL 5 has added support for converting additional data types.
Analyzedb: Converting datetime to a string for correct comparison with the value
saved in the file.
el8_migrate_localte.py, gparray.py, gpcatalog.py and gpcheckcat: using the Bool
type instead of comparing with a string.
gpcheckcat, repair_missing_extraneous.py and unique_index_violation_check:
using python list instead of string parsing.
2. PyGreSQL 5 added support for closing a connection when using the with
construct. Because of this, in a number of places, reading from the cursor took
place after the connection was closed.
3. PyGreSQL 5 does not end the transaction if an error occurs, which leads to a
possible connection leak if an error occurs in the connect function. So catch
errors that happen in the connect function.
4. Add closure of the connection saved in context after the scenario in behave
tests.
5. Add closure to the connection if it does not return from the function.
6. Use the python wrapper for the connect function instead of C one.
7. Use a custom cursor to disable row postprocessing to avoid correcting a large
amount of code.
8. Fix the bool and array format in isolation2 tests.
9. Add notifications processing to isolation2 tests.
10. Also fix the notifications processing in the resgroup_query_mem test.
11. Fix the notifications processing in gpload.
12. Fix pg_config search when building deb packages.