Avoid using gp_tablespace_segment_location in a correlated subplan (#2138)
The gp_tablespace_segment_location function is marked as being executed on all
segments. When used in a correlated subplan, when its argument depends on the
outer plan, it may produce incorrect results. We plan to disable such plans.
Rewrite arenadata_toolkit to avoid using the correlated function on segments in
the subplan.

Ticket: ADBDEV-8931