Extend shared tuplestore lifetime on EntryDB (#148)
It was previously assumed that on EntryDB the shared tuplestore does not need
the full transaction lifetime, since it doesn't have to last between queries.
However, due to the new order of resource cleanup, executor lifetime is not
enough anymore. Now the shared tuplestore requires full transaction-level
MemoryContext and ResourceOwner, even on the coordinator.

Ticket: ADBDEV-8790