Fix orphaned segment files for added and rolled back columns for AOCS tables (#173)
Problem description:
In case we added a new column to a AOCS table in a transaction that was later
rolled back, a non-empty segment file could be left, and vacuum didn't truncate
it.

Root cause:
Vacuum does truncation of AOCS segment files in 2 places:
1. In AOCSSegmentFileTruncateToEOF(). It iterated over 'vpinfo.nEntry', which
doesn't count the added and rolled back attribute, so its segment file was not
processed here.
2. In AOCSCompaction_DropSegmentFile() function (if compaction is required for
the segment). It iterates over attributes retrieved by
RelationGetNumberOfAttributes(). But, the added and rolled back attribute is not
counted by RelationGetNumberOfAttributes(), so its segment file was not
processed here as well.

Fix:
In AOCSSegmentFileTruncateToEOF() iterate and truncate all segment files for
a given 'segno' until they exist. We stop once we didn't find a file and column
number is above 'vpinfo.nEntry'. We assume that the segment files exist or stop
existing for all columns thereafter - similar logic we already use in
ao_foreach_extent_file().

We do it not in AOCSCompaction_DropSegmentFile(), as it will not be called if
compaction is not required, while the orphaned segment files may still exist in
this case.

Changes from original commit:
We cannot use 'attnum' to iterate through segment files, since in GPDB 7 the
filenumber is stored explicitly in 'pg_attribute_encoding' system table.
Changes to this table are also rolled back, so we won't be able to find an
entry for the 'attnum' in there. Instead, iterate through possible filenums
and find the corresponding 'attnum'. If the file exists but has no entry in
'pg_attribute_encoding', it's safe to drop.
Notably, we can only stop when both filenum and filenum+1600 are missing,
since table rewrites can replace filenum with filenum+1600. This means we
must check filenums in order 1, 1601, 2, 1602, 3, 1603 etc. to ensure we
don't miss any.
Since filenums are allocated in order, for N-column table it's enough to
check filenums up to N and N+1600, and then more only if they exist.

(cherry picked from commit b5f63c6571e4300187a1d88e266520d0187fdd82)

Co-authored-by: Maxim Michkov <m.michkov@arenadata.io>