Fix Orca CCTETest unit test (#162)
The patch fixes a test added in b1b31a2 that was broken because of the
cost model changes related to hash join introduced in 55a0fcb.
As a result of these changes, both hash joins present in the test
are expected to have a higher cost.

The fix for the issue is a simple update of the plan cost reported by the test
that makes it comply with the current cost model.

(cherry picked from commit 5d7693ec9f9f19c5f76a4b93ca09bd6037f570c3)